src/llm/requirements.txt
# Dependências Python para o serviço LLM
# Usado pelo Agente Key para geração de conteúdo

# Dependências básicas
python-dotenv>=1.0.0

# Google Gemini API
google-generativeai>=0.3.0

# HTTP requests (se necessário para scripts Python)
requests>=2.25.1

# Yahoo Finance (para AgentJulia - opcional)
yfinance>=0.2.0
pandas>=1.5.0

